GEECS-Console 0.15.1: main_window slimming step 1 — HealthPoller + tooltips extracted (#534)
